replace old examples with newer variants
