+++ examples/mnist/spark/mnist_spark.py
@@ -15,7 +15,7 @@ import tensorflow as tf
 from datetime import datetime
 
 from tensorflowonspark import TFCluster
-import mnist_dist
+import mnist_dist_dataset
 
 sc = SparkContext(conf=SparkConf().setAppName("mnist_spark"))
 executors = sc._conf.get("spark.executor.instances")
@@ -25,7 +25,6 @@ num_ps = 1
 parser = argparse.ArgumentParser()
 parser.add_argument("--batch_size", help="number of records per batch", type=int, default=100)
 parser.add_argument("--epochs", help="number of epochs", type=int, default=1)
-parser.add_argument("--export_dir", help="HDFS path to export saved_model", default="mnist_export")
 parser.add_argument("--format", help="example format: (csv|tfr)", choices=["csv", "tfr"], default="csv")
 parser.add_argument("--images", help="HDFS path to MNIST images in parallelized format")
 parser.add_argument("--labels", help="HDFS path to MNIST labels in parallelized format")
@@ -56,23 +55,18 @@ if args.format == "tfr":
     return (image, label)
 
   dataRDD = images.map(lambda x: toNumpy(bytes(x[0])))
-else:  # "csv"
-  print("zipping images and labels")
-  # If partitions of images/labels don't match, you can use the following code:
-  # images = sc.textFile(args.images).map(lambda ln: [int(x) for x in ln.split(',')]).zipWithIndex().map(lambda x: (x[1], x[0]))
-  # labels = sc.textFile(args.labels).map(lambda ln: [float(x) for x in ln.split(',')]).zipWithIndex().map(lambda x: (x[1], x[0]))
-  # dataRDD = images.join(labels).map(lambda x: (x[1][0], x[1][1]))
+else:  # args.format == "csv":
   images = sc.textFile(args.images).map(lambda ln: [int(x) for x in ln.split(',')])
   labels = sc.textFile(args.labels).map(lambda ln: [float(x) for x in ln.split(',')])
+  print("zipping images and labels")
   dataRDD = images.zip(labels)
 
-cluster = TFCluster.run(sc, mnist_dist.map_fun, args, args.cluster_size, num_ps, args.tensorboard, TFCluster.InputMode.SPARK, log_dir=args.model)
+cluster = TFCluster.run(sc, mnist_dist_dataset.map_fun, args, args.cluster_size, num_ps, args.tensorboard, TFCluster.InputMode.SPARK)
 if args.mode == "train":
   cluster.train(dataRDD, args.epochs)
-else:  # inference
-  predRDD = cluster.inference(dataRDD)
-  predRDD.saveAsTextFile(args.output)
-
-cluster.shutdown(grace_secs=30)
+else:
+  labelRDD = cluster.inference(dataRDD)
+  labelRDD.saveAsTextFile(args.output)
+cluster.shutdown()
 
 print("{0} ===== Stop".format(datetime.now().isoformat()))
