--- examples/mnist/spark/mnist_spark_dataset.py
@@ -1,72 +1,0 @@
-# Copyright 2017 Yahoo Inc.
-# Licensed under the terms of the Apache 2.0 license.
-# Please see LICENSE file in the project root for terms.
-
-from __future__ import absolute_import
-from __future__ import division
-from __future__ import print_function
-
-from pyspark.context import SparkContext
-from pyspark.conf import SparkConf
-
-import argparse
-import numpy
-import tensorflow as tf
-from datetime import datetime
-
-from tensorflowonspark import TFCluster
-import mnist_dist_dataset
-
-sc = SparkContext(conf=SparkConf().setAppName("mnist_spark"))
-executors = sc._conf.get("spark.executor.instances")
-num_executors = int(executors) if executors is not None else 1
-num_ps = 1
-
-parser = argparse.ArgumentParser()
-parser.add_argument("--batch_size", help="number of records per batch", type=int, default=100)
-parser.add_argument("--epochs", help="number of epochs", type=int, default=1)
-parser.add_argument("--format", help="example format: (csv|tfr)", choices=["csv", "tfr"], default="csv")
-parser.add_argument("--images", help="HDFS path to MNIST images in parallelized format")
-parser.add_argument("--labels", help="HDFS path to MNIST labels in parallelized format")
-parser.add_argument("--model", help="HDFS path to save/load model during train/inference", default="mnist_model")
-parser.add_argument("--cluster_size", help="number of nodes in the cluster", type=int, default=num_executors)
-parser.add_argument("--output", help="HDFS path to save test/inference output", default="predictions")
-parser.add_argument("--readers", help="number of reader/enqueue threads", type=int, default=1)
-parser.add_argument("--steps", help="maximum number of steps", type=int, default=1000)
-parser.add_argument("--tensorboard", help="launch tensorboard process", action="store_true")
-parser.add_argument("--mode", help="train|inference", default="train")
-parser.add_argument("--rdma", help="use rdma connection", default=False)
-args = parser.parse_args()
-print("args:", args)
-
-print("{0} ===== Start".format(datetime.now().isoformat()))
-
-if args.format == "tfr":
-  images = sc.newAPIHadoopFile(args.images, "org.tensorflow.hadoop.io.TFRecordFileInputFormat",
-                               keyClass="org.apache.hadoop.io.BytesWritable",
-                               valueClass="org.apache.hadoop.io.NullWritable")
-
-  def toNumpy(bytestr):
-    example = tf.train.Example()
-    example.ParseFromString(bytestr)
-    features = example.features.feature
-    image = numpy.array(features['image'].int64_list.value)
-    label = numpy.array(features['label'].int64_list.value)
-    return (image, label)
-
-  dataRDD = images.map(lambda x: toNumpy(bytes(x[0])))
-else:  # args.format == "csv":
-  images = sc.textFile(args.images).map(lambda ln: [int(x) for x in ln.split(',')])
-  labels = sc.textFile(args.labels).map(lambda ln: [float(x) for x in ln.split(',')])
-  print("zipping images and labels")
-  dataRDD = images.zip(labels)
-
-cluster = TFCluster.run(sc, mnist_dist_dataset.map_fun, args, args.cluster_size, num_ps, args.tensorboard, TFCluster.InputMode.SPARK)
-if args.mode == "train":
-  cluster.train(dataRDD, args.epochs)
-else:
-  labelRDD = cluster.inference(dataRDD)
-  labelRDD.saveAsTextFile(args.output)
-cluster.shutdown()
-
-print("{0} ===== Stop".format(datetime.now().isoformat()))
